code fixes

+++ 2-do_deploy_web_static.py
@@ -18,11 +18,15 @@ def do_pack():
     print("Packing web_static to {}".format(archive_path))
 
     local("mkdir -p versions")
-    local("tar -cvzf {} web_static".format(archive_path))
+    result = local("tar -cvzf {} web_static".format(archive_path))
+
+    if result.failed:
+        return None
 
     print("Successfully packed web_static to {}".format(archive_path))
 
     return archive_path
+
 
 def do_deploy(archive_path):
     """ Distributes an archive to webserver """
@@ -38,7 +42,11 @@ def do_deploy(archive_path):
 
     put(archive_path, "/tmp/")
     run("mkdir -p {}".format(dir_path))
-    run("tar -xzf /tmp/{} -C {}".format(archive_name, dir_path))
+    result = run("tar -xzf /tmp/{} -C {}".format(archive_name, dir_path))
+
+    if result.failed:
+        return False
+
     run("cp -r {}/web_static/* {}".format(dir_path, dir_path))
     run("rm -rf /tmp/{} {}/web_static".format(archive_name, dir_path))
     run("rm -rf /data/web_static/current")
